First pass at implementing an API for use by LCLS
This commit adds cspad support as a first test. FormatXTCCspadExtern overrides FormatXTCCspad to allow in-memory reading of the xtc stream without using a locator file so that the calling program can be in charge of reading the psana events. The main class PsanaEventProcessor invokes the dials.stills_process Processor to do the work.

exercise_api demos how to use the API.